+++ releases.md
@@ -1,3 +1,7 @@
 # Releases
 
+## Unreleased
+
+  - Allow `Protocol::Multipart::Parser` to consume chunk-readable inputs in addition to IO streams.
+
 ## v0.1.0